+++ tests/common/dummy_data.rs
@@ -43,32 +43,45 @@ pub enum DummyImage {
 
 #[derive(Clone)]
 pub struct DummyData {
+    pub project_alpha: DummyProjectAlpha,
+    pub project_beta: DummyProjectBeta,
+    pub organization_zeta: DummyOrganizationZeta,
+}
+
+#[derive(Clone)]
+pub struct DummyProjectAlpha {
     // Alpha project:
     // This is a dummy project created by USER user.
     // It's approved, listed, and visible to the public.
-    pub alpha_project_id: String,
-    pub alpha_project_slug: String,
-    pub alpha_version_id: String,
-    pub alpha_thread_id: String,
-    pub alpha_file_hash: String,
-    pub alpha_team_id: String,
+    pub project_id: String,
+    pub project_slug: String,
+    pub version_id: String,
+    pub thread_id: String,
+    pub file_hash: String,
+    pub team_id: String,
+}
 
+#[derive(Clone)]
+pub struct DummyProjectBeta {
     // Beta project:
     // This is a dummy project created by USER user.
     // It's not approved, unlisted, and not visible to the public.
-    pub beta_project_id: String,
-    pub beta_project_slug: String,
-    pub beta_version_id: String,
-    pub beta_thread_id: String,
-    pub beta_file_hash: String,
-    pub beta_team_id: String,
+    pub project_id: String,
+    pub project_slug: String,
+    pub version_id: String,
+    pub thread_id: String,
+    pub file_hash: String,
+    pub team_id: String,
+}
 
+#[derive(Clone)]
+pub struct DummyOrganizationZeta {
     // Zeta organization:
     // This is a dummy organization created by USER user.
     // There are no projects in it.
-    pub zeta_organization_id: String,
-    pub zeta_organization_title: String,
-    pub zeta_team_id: String,
+    pub organization_id: String,
+    pub organization_title: String,
+    pub team_id: String,
 }
 
 pub async fn add_dummy_data(test_env: &TestEnvironment) -> DummyData {
@@ -88,34 +101,36 @@ pub async fn add_dummy_data(test_env: &TestEnvironment) -> DummyData {
 
     let zeta_organization = add_organization_zeta(test_env).await;
 
-    sqlx::query("INSERT INTO dummy_data (version) VALUES ($1)")
+    sqlx::query("INSERT INTO dummy_data (update_id) VALUES ($1)")
         .bind(DUMMY_DATA_UPDATE)
         .execute(pool)
         .await
         .unwrap();
 
     DummyData {
-        alpha_team_id: alpha_project.team.to_string(),
-        beta_team_id: beta_project.team.to_string(),
+        project_alpha: DummyProjectAlpha {
+            team_id: alpha_project.team.to_string(),
+            project_id: alpha_project.id.to_string(),
+            project_slug: alpha_project.slug.unwrap(),
+            version_id: alpha_version.id.to_string(),
+            thread_id: alpha_project.thread_id.to_string(),
+            file_hash: alpha_version.files[0].hashes["sha1"].clone(),
+        },
 
-        alpha_project_id: alpha_project.id.to_string(),
-        beta_project_id: beta_project.id.to_string(),
+        project_beta: DummyProjectBeta {
+            team_id: beta_project.team.to_string(),
+            project_id: beta_project.id.to_string(),
+            project_slug: beta_project.slug.unwrap(),
+            version_id: beta_version.id.to_string(),
+            thread_id: beta_project.thread_id.to_string(),
+            file_hash: beta_version.files[0].hashes["sha1"].clone(),
+        },
 
-        alpha_project_slug: alpha_project.slug.unwrap(),
-        beta_project_slug: beta_project.slug.unwrap(),
-
-        alpha_version_id: alpha_version.id.to_string(),
-        beta_version_id: beta_version.id.to_string(),
-
-        alpha_thread_id: alpha_project.thread_id.to_string(),
-        beta_thread_id: beta_project.thread_id.to_string(),
-
-        alpha_file_hash: alpha_version.files[0].hashes["sha1"].clone(),
-        beta_file_hash: beta_version.files[0].hashes["sha1"].clone(),
-
-        zeta_organization_id: zeta_organization.id.to_string(),
-        zeta_team_id: zeta_organization.team_id.to_string(),
-        zeta_organization_title: zeta_organization.title,
+        organization_zeta: DummyOrganizationZeta {
+            organization_id: zeta_organization.id.to_string(),
+            team_id: zeta_organization.team_id.to_string(),
+            organization_title: zeta_organization.title,
+        },
     }
 }
 
@@ -125,27 +140,29 @@ pub async fn get_dummy_data(test_env: &TestEnvironment) -> DummyData {
 
     let zeta_organization = get_organization_zeta(test_env).await;
     DummyData {
-        alpha_team_id: alpha_project.team.to_string(),
-        beta_team_id: beta_project.team.to_string(),
+        project_alpha: DummyProjectAlpha {
+            team_id: alpha_project.team.to_string(),
+            project_id: alpha_project.id.to_string(),
+            project_slug: alpha_project.slug.unwrap(),
+            version_id: alpha_version.id.to_string(),
+            thread_id: alpha_project.thread_id.to_string(),
+            file_hash: alpha_version.files[0].hashes["sha1"].clone(),
+        },
 
-        alpha_project_id: alpha_project.id.to_string(),
-        beta_project_id: beta_project.id.to_string(),
+        project_beta: DummyProjectBeta {
+            team_id: beta_project.team.to_string(),
+            project_id: beta_project.id.to_string(),
+            project_slug: beta_project.slug.unwrap(),
+            version_id: beta_version.id.to_string(),
+            thread_id: beta_project.thread_id.to_string(),
+            file_hash: beta_version.files[0].hashes["sha1"].clone(),
+        },
 
-        alpha_project_slug: alpha_project.slug.unwrap(),
-        beta_project_slug: beta_project.slug.unwrap(),
-
-        alpha_version_id: alpha_version.id.to_string(),
-        beta_version_id: beta_version.id.to_string(),
-
-        alpha_thread_id: alpha_project.thread_id.to_string(),
-        beta_thread_id: beta_project.thread_id.to_string(),
-
-        alpha_file_hash: alpha_version.files[0].hashes["sha1"].clone(),
-        beta_file_hash: beta_version.files[0].hashes["sha1"].clone(),
-
-        zeta_organization_id: zeta_organization.id.to_string(),
-        zeta_team_id: zeta_organization.team_id.to_string(),
-        zeta_organization_title: zeta_organization.title,
+        organization_zeta: DummyOrganizationZeta {
+            organization_id: zeta_organization.id.to_string(),
+            team_id: zeta_organization.team_id.to_string(),
+            organization_title: zeta_organization.title,
+        },
     }
 }
